Вносит правки
