+++ js/api.js
@@ -1,19 +1,20 @@
 const GET_URL = 'https://28.javascript.pages.academy/kekstagram/data';
 const SEND_URL = 'https://28.javascript.pages.academy/kekstagram';
+const GET_ERROR_TEXT = 'Ошибка загрузки данных. Попробуйте перезагрузить страницу';
+const SEND_ERROR_TEXT = 'Ошибка отправки данных. Попробуйте еще раз';
 
 export const getData = (onSuccess, onError) => fetch(
   GET_URL)
   .then((response) => {
     if (!response.ok) {
-      throw new Error('Ошибка загрузки данных. Пожалуйста перезагрузите страницу');
+      throw new Error();
     }
+
     return response.json();
   })
-  .then((postsData) => {
-    onSuccess(postsData);
-  })
+  .then(onSuccess)
   .catch(() => {
-    onError('Ошибка загрузки данных. Попробуйте перезагрузить страницу');
+    onError(GET_ERROR_TEXT);
   });
 
 export const sendData = (body, onSuccess, onError) => fetch(
@@ -24,8 +25,9 @@ export const sendData = (body, onSuccess, onError) => fetch(
   })
   .then((response) => {
     if (!response.ok) {
-      throw new Error('Ошибка отправки данных. Попробуйте еще раз');
+      throw new Error(SEND_ERROR_TEXT);
     }
+
     onSuccess();
   })
   .catch(() => {
