+++ js/main.js
@@ -1,7 +1,7 @@
 import {onGetPostSuccess} from './create-post.js';
 import './form.js';
 import {getData} from './api.js';
-import {onGetPostError} from './utils.js';
+import {onGetPostError} from './dialog.js';
 
 getData(onGetPostSuccess, onGetPostError);
