+++ js/utils.js
@@ -1,5 +1,3 @@
-const ALERT_SHOW_TIME = 5000;
-
 export const isStringOptimal = (str, maxLength) => str.length <= maxLength;
 
 isStringOptimal('check', 3);
@@ -46,24 +44,3 @@ export const getRandomInteger = (a, b) => {
 export const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
 
 export const isEscapeKey = (key) => key === 'Escape';
-
-export const onGetPostError = (message) => {
-  const alertContainer = document.createElement('div');
-  alertContainer.style.zIndex = '100';
-  alertContainer.style.position = 'absolute';
-  alertContainer.style.left = '0';
-  alertContainer.style.top = '0';
-  alertContainer.style.right = '0';
-  alertContainer.style.padding = '10px 3px';
-  alertContainer.style.fontSize = '30px';
-  alertContainer.style.textAlign = 'center';
-  alertContainer.style.backgroundColor = 'red';
-
-  alertContainer.textContent = message;
-
-  document.body.append(alertContainer);
-
-  setTimeout(() => {
-    alertContainer.remove();
-  }, ALERT_SHOW_TIME);
-};
